pyproject.toml
[tool.poetry]
name = "berries-backend"
version = "0.1.0"
description = "Backend for the Berries Financial Assistant"
authors = ["Your Name <amir224030@bue.edu.eg>"]
readme = "README.md"
packages = [{ include = "rag_pipeline", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0.0"
python-dotenv = "^1.1.0"
alpaca-py = "^0.39.2"
yfinance = "^0.2.55"
langchain = "^0.3.23"
langchain-community = "^0.3.21"
langchain-huggingface = "^0.1.2"
pdfminer-six = "^20250327"
chromadb = "<0.7.0"
unstructured = {extras = ["pytesseract"], version = "^0.17.2"}
pi-heif = "^0.22.0"
unstructured-inference = "^0.8.10"
pdf2image = "^1.17.0"
pytesseract = "^0.3.13"
pdfplumber = "^0.11.6"
langchain-google-genai = "^2.1.2"
langchain-chroma = "^0.2.2"
stockstats = "^0.6.4"
numpy = "^2.2.4"
pandas = "^2.2.3"
matplotlib = "^3.10.1"
torch = "^2.6.0"
finrl = "^0.3.7"
gym = "^0.26.2"
stable-baselines3 = "^2.6.0"
alpaca-trade-api = "^3.2.0"
exchange-calendars = "^4.10"
wrds = "^3.3.0"


[tool.poetry.scripts]
app = "rag_pipeline.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
